Relocate hydrateTaskFromEvent
So we can group it with the useSubmit (since the hook uses it).
